+++ tests/testthat/test_plot.functions.R
@@ -40,37 +40,35 @@ modellist <- list(linear, emax, emax.class, emax.noplac)
 
 test_that("plot.mbnma.network functions correctly", {
 
-  expect_silent(plot(network, layout_in_circle = TRUE,
+  expect_silent(plot(network, layout = igraph::as_star(),
                                edge.scale=1, label.distance=0))
 
-  expect_silent(plot(network, layout_in_circle = FALSE,
+  expect_silent(plot(network, layout = igraph::with_fr(),
                                edge.scale=1, label.distance=0))
 
-  expect_silent(plot(network, layout_in_circle = FALSE,
+  expect_silent(plot(network, layout = igraph::in_circle(),
                                edge.scale=0.5, label.distance=10))
 
-  g1 <- plot(network, layout_in_circle = TRUE,
-             level="treatment")
-  g2 <- plot(network, layout_in_circle = TRUE,
-             level="agent")
+  g1 <- plot(network, level="treatment")
+  g2 <- plot(network, level="agent")
 
   expect_silent(plot(g1))
   expect_silent(plot(g2))
 
   expect_equal(length(igraph::V(g1))==length(igraph::V(g2)), FALSE)
 
-  expect_silent(plot(network, layout_in_circle = TRUE,
+  expect_silent(plot(network, layout = igraph::in_circle(),
                      level="agent", remove.loops = TRUE))
 
-  expect_warning(plot(net.noplac, layout_in_circle = TRUE,
+  expect_warning(plot(net.noplac, layout=igraph::as_star(),
                      level="agent"))
 
-  expect_warning(plot(net.noplac, layout_in_circle = TRUE,
+  expect_warning(plot(net.noplac, layout=igraph::with_fr(),
                       level="agent", doselink = 5))
 
-  g1 <- plot(network, layout_in_circle = TRUE,
+  g1 <- plot(network,
             level="treatment", v.color = "agent")
-  g2 <- plot(net.noplac, layout_in_circle = TRUE,
+  g2 <- plot(net.noplac,
              level="treatment", v.color="agent", doselink = 1)
 
   expect_equal("Placebo_0" %in% names(igraph::V(g1)), TRUE)
@@ -82,7 +80,7 @@ test_that("plot.mbnma.network functions correctly", {
   expect_equal(length(unique(igraph::V(g2)$color)), length(net.noplac$agents)+1)
   expect_equal(length(unique(igraph::E(g2)$color)), 2)
 
-  expect_error(plot(network, layout_in_circle = TRUE,
+  expect_error(plot(network, layout=igraph::in_circle(),
                               level="class"))
 
 })